# public/api/old_swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Modasti Backend API",
    "termsOfService": "",
    "contact": {
      "email": "biz@riseapps.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "stage-api.modasti.net",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/getContests": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "get contests",
        "description": "",
        "operationId": "getContests",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Contests"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getContestPhotos": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "get Contest Photos",
        "description": "",
        "operationId": "getContestPhotos",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Contest Photos",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "contestId": {
                  "type": "number",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "$ref": "#/definitions/ContestPhotos"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/publishContestPhoto": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "publish contest photo",
        "description": "",
        "operationId": "publishContestPhoto",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "uploadPhoto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "contestId": {
                  "type": "integer",
                  "required": true
                },
                "imageName": {
                  "type": "string",
                  "required": "true"
                },
                "image": {
                  "type": "string",
                  "required": "true"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_token",
      "in": "query"
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "fname": {
          "type": "string"
        },
        "lname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "user_type": {
          "type": "string"
        },
        "following_counter": {
          "type": "number"
        },
        "follower_counter": {
          "type": "number"
        },
        "is_followed": {
          "type": "boolean"
        },
        "photo": {
          "type": "object",
          "properties": {
            "table_id": {
              "type": "integer"
            },
            "photo_name": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "Contests": {
      "type": "array",
      "items": {
        "type": "object",
        "$ref": "#/definitions/Contest"
      }
    },
    "Contest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title_en": {
          "type": "string"
        },
        "text_en": {
          "type": "string"
        },
        "text2_en": {
          "type": "string"
        },
        "created": {
          "type": "string"
        },
        "expires": {
          "type": "string"
        },
        "time_diff": {
          "type": "string"
        },
        "is_photo_submitted": {
          "type": "boolean"
        },
        "photo": {
          "type": "object",
          "properties": {
            "table_id": {
              "type": "integer"
            },
            "photo_name": {
              "type": "string"
            }
          }
        },
        "winners": {
          "$ref": "#/definitions/Winners"
        }
      }
    },
...
    "ContestPhotos": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ContestPhoto"
      }
    },
    "ContestPhoto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "contest_id": {
          "type": "integer"
        },
        "photo_id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "likes": {
          "type": "integer"
        },
        "date_created": {
          "type": "string"
        },
        "date_likes": {
          "type": "string"
        },
        "is_liked": {
          "type": "boolean"
        },
        "photo": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "photo_name": {
              "type": "string"
            }
          }
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "Winners": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Winner"
      }
    },
    "Winner": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "contest_id": {
          "type": "integer"
        },
        "photo_id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "contest_title": {
          "type": "string"
        },
        "win_place": {
          "type": "integer"
        },
        "photo": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "photo_name": {
              "type": "string"
            }
          }
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    }
  }
}